+++ bin/bwa_mem.pl
@@ -313,7 +313,8 @@ This also impacts the number of threads used by BWA mapping steps.
 =item B<-fragment>
 
 Split input into fragements of X million repairs.  To prevent variability in data processing either
-set this to a very large number or ensure that it is not changed.
+set this to a very large number or ensure that it is not changed.  Values >= 5000 indicate that data
+should not be split.
 
 =item B<-nomarkdup>
 
@@ -365,10 +366,10 @@ Mark reads as QCFAIL (0x200, 512) using the mismatchQc program, also adds aux ta
 WARNING:
 bwa_mem.pl will exclude all QCFAIL reads from mapping. If a BAM/CRAM file has been created using
 this option please ensure that you pre-process the file to remove the flag 512 if you intend to
-reprocess based on that output. 
+reprocess based on that output.
 
-The script mmFlagModifier -m (--remove) can process a bam file to remove any occurences of 
-flag 512 where the read also has the tag mm:A:Y . 
+The script mmFlagModifier -m (--remove) can process a bam file to remove any occurences of
+flag 512 where the read also has the tag mm:A:Y .
 
 e.g.