+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 4.3.2
 
+* Correct mem free issue when writing CRAM in mismatchQc
 * Resolved #40 - faster recovery times
 * Resolved #39 - handle splitting correctly